+++ R/map_layer_arc.R
@@ -20,8 +20,10 @@ mapdeckArcDependency <- function() {
 #' @param layer_id single value specifying an id for the layer. Use this value to
 #' distinguish between shape layers of the same type. Layers with the same id are likely
 #' to conflict and not plot correctly
-#' @param origin vector of longitude and latitude columns, or an \code{sfc} column
-#' @param destination vector of longitude and latitude columns, or an \code{sfc} column
+#' @param origin vector of longitude and latitude columns, and optionally an elevation column,
+#' or an \code{sfc} column
+#' @param destination vector of longitude and latitude columns, and optionally an elevatino column,
+#' or an \code{sfc} column
 #' @param id an id value in \code{data} to identify layers when interacting in Shiny apps.
 #' @param stroke_from column of \code{data} or hex colour to use as the staring stroke colour.
 #' If using a hex colour, use either a single value, or a vector the same length as \code{data}
@@ -286,7 +288,7 @@ add_arc <- function(
 		geometry_column <- c( "origin", "destination" )
 		shape <- rcpp_od_geojson( data, l, geometry_column, digits, "arc" )
   } else if ( tp == "df" ) {
-  	geometry_column <- list( origin = c("start_lon", "start_lat"), destination = c("end_lon", "end_lat") )
+  	geometry_column <- list( origin = c("start_lon", "start_lat", "start_elev"), destination = c("end_lon", "end_lat", "end_elev") )
   	shape <- rcpp_od_geojson_df( data, l, geometry_column, digits, "arc" )
   } else if ( tp == "sfencoded" ) {
   	geometry_column <- c("origin", "destination")