+++ inst/htmlwidgets/mapdeck_location.js
@@ -3,10 +3,12 @@ function md_change_location( map_id, map_type, location, zoom, pitch, bearing, d
   var currentLon, currentLat, currentPitch, currentBearing, currentZoom;
 
   if( map_type == "google_map" ) {
-  	console.log( location );
+  	//console.log( location );
   	window[map_id + 'map'].setCenter( { lat: location[1], lng: location[0] } );
   	window[map_id + 'map'].setZoom( zoom );
   } else {
+
+  	console.log( window[ map_id + 'map' ].viewState );
 
 	  if ( window[ map_id + 'map'].viewState["default-view"] !== undefined ) {
 	  	currentLon = location === null ? window[ map_id + 'map'].viewState["default-view"].longitude : location[0];
@@ -22,17 +24,21 @@ function md_change_location( map_id, map_type, location, zoom, pitch, bearing, d
 	    currentZoom = zoom === null ? window[ map_id + 'map'].viewState.zoom : zoom;
 	  }
 
+	  console.log( currentLon );
+
 		window[map_id + 'map'].setProps({
-	    viewState: {
+	    initialViewState: {
 	      longitude: currentLon,
 	      latitude: currentLat,
 	      zoom: currentZoom,
 	      pitch: currentPitch,
 	      bearing: currentBearing,
 	      transitionInterpolator: transition === "fly" ? new deck.FlyToInterpolator() : new deck.LinearInterpolator(),
-	      transitionDuration: duration
-	    },
+	      transitionDuration: duration,
+	      controller: true
+	    }
 	  });
+
   }
 }