+++ NEWS.md
@@ -1,6 +1,9 @@
 # mapdeck 0.3
 
 
+* v8.* of Deck.gl
+	- Arc layer supports 3D positions
+* `show_view_state` option to overlay the current View State [issue238](https://github.com/SymbolixAU/mapdeck/issues/238)
 * performance improvements for dates and factors [issue235](https://github.com/SymbolixAU/mapdeck/issues/235)
 * `add_bitmap()` to add a bitmap / image to the map
 * input$map_view_change now observed [issue211](https://github.com/SymbolixAU/mapdeck/issues/211) - thanks @zacdav